requirements.txt
-i http://pypi.doubanio.com/simple
django == 1.11.21
Markdown == 3.1.1
django-stdimage==3.2.0
PyMySQL==0.9.3
mistune